+++ jam/commands/ScalersCmd.java
@@ -51,8 +51,7 @@ public final class ScalersCmd extends AbstractCommand {
 	 * If gets used as an Action, i.e., gets passed null, defaults to READ.
 	 */
 	protected void execute(final Object[] cmdParams) {
-		final int param = (null == cmdParams) ? READ : ((Integer) cmdParams[0])
-				.intValue();
+		final int param = (null == cmdParams) ? READ : (Integer) cmdParams[0];
 		if (param == READ) {
 			readScalers();
 		} else if (param == ZERO) {
@@ -67,9 +66,9 @@ public final class ScalersCmd extends AbstractCommand {
 	protected void executeParse(final String[] cmdTokens) {
 		final Object[] params = new Object[1];
 		if (cmdTokens[0].equals("read")) {
-			params[0] = Integer.valueOf(READ);
+			params[0] = READ;
 		} else if (cmdTokens[0].equals("zero")) {
-			params[0] = Integer.valueOf(ZERO);
+			params[0] = ZERO;
 		}
 		execute(params);
 	}
